Add test for known issue in library
